worker/requirements.txt
supabase>=2.0.0
pandas>=2.0.0
numpy>=1.24.0
sentence-transformers>=2.2.0
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
scikit-learn>=1.3.0
